requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
Django==5.0.7
django-appconf==1.0.6
django-browser-reload==1.13.0
django-compressor==4.5.1
django-cors-headers==4.4.0
django-filter==24.3
django-libsass==0.9
django-restframework==0.0.1
django-tailwind==3.8.0
djangorestframework==3.15.2
idna==3.7
Jinja2==3.1.4
libsass==0.23.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
pandas==2.2.2
pillow==10.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
rcssmin==1.1.2
requests==2.32.3
rich==13.7.1
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
